librbd, cephdriver: move to new librbd.Image struct which encapsulates image operations for a specific image
